Convert all JAX-RS classes to Spring controllers
This will allow us to remove the dependency on Jersey and rely on
Spring boot to easily start a local server in development without
the need to provide a Tomcat environment.

Moreover, Spring was already used sparsely for some actions so to
be consistent, let's rather rely only on one framework.