Remove all DAO and rely on Spring repositories
Rather than defining a method for each kind of query and create
the boilerplate code to create a query and execute it, let's rather
rely on Spring Data repositories to have a more DRY code and also
more efficient as the previous code as we were doing string
concatenation to produce JPQL requests.

Some methods have been refined as well to be more DRY (a few methods
share like 90% of their code apart from a condition or a loop ; this
commit tries to refactor that as much as possible).