Get rid of the manual XML generation and parsing
Rely on Jackson to ease code factoring for the serialization and
de-serialization of entities.

For now, XML is only really supported but all the necessary stuff
is present to slowly introduce JSON and ease the front-end and
gain some bandwidth if needed.

Also, relying on Jackson has widely improved the general readability
of the code and will certainly bring a performance boost overall.