Properly build portfolio documents
This may not exactly produce the same output as previously but at
least, we are using the same process throughout the application
code and we can tweak this generation process if necessary later.

Down the line, we want to deal with objects to be able to pull some
specific attributes. We don't care about strings even though they
are what's being generated; that's Jackson responsibility.