Bring the basis of custom SpEL security expressions
Thus, we can query our own authorization system through SpEL expressions
passed to the security annotations.

This avoids passing the user's id to every single service method that
needs authorization checks and we lessen the use of the code smell
brought by the `BusinessException` thing.

Plus we get the right HTTP code out of the box when the user is not
authorized to trigger a method.