Rework and test the import process of portfolios
Totally avoid relying on the disk for processing the zip file. Since
we don't need it and we need to load content in memory anyway, we
can save some time and resources, we lower the pressure on disk and
avoid any pitfalls cleaning the temporary files.

Also, fix some disparities with the implementation from version 2.